refactor: extract undo operations module
